pyproject.toml
[tool.poetry]
name = "nonebot_plugin_chatrecorder"
version = "0.6.2"
description = "适用于 Nonebot2 的聊天记录插件"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-chatrecorder"
repository = "https://github.com/noneplugin/nonebot-plugin-chatrecorder"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.3.0"
nonebot-plugin-orm = ">=0.7.0,<1.0.0"
nonebot-plugin-uninfo = ">=0.6.0,<1.0.0"
nonebot-plugin-localstore = ">=0.6.0,<1.0.0"

[tool.poetry.group.dev.dependencies]
nonebot2 = { version = "^2.3.0", extras = ["fastapi"] }
nonebot-plugin-orm = { version = ">=0.7.0,<1.0.0", extras = ["default"] }
nonebot-plugin-datastore = "^1.0.0"

[tool.poetry.group.test.dependencies]
nonebug = "^0.4.2"
pytest-cov = "^5.0.0"
pytest-xdist = "^3.0.2"
pytest-asyncio = "^0.24.0"
asyncpg = "^0.30.0"
aiomysql = "^0.2.0"

[tool.poetry.group.adapters]
optional = true

[tool.poetry.group.adapters.dependencies]
nonebot-adapter-onebot = "^2.4.4"
nonebot-adapter-console = "^0.6.0"
nonebot-adapter-kaiheila = "^0.3.4"
nonebot-adapter-telegram = "^0.1.0b17"
nonebot-adapter-feishu = "^2.6.2"
nonebot-adapter-discord = "^0.1.8"
nonebot-adapter-dodo = "^0.2.1"
nonebot-adapter-satori = "^0.12.3"
nonebot-adapter-qq = "^1.5.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.nonebot]
plugins = ["nonebot_plugin_chatrecorder"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "Q"]
ignore = ["E402", "E501", "E711", "C901", "UP037"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "@overload",
  "except ImportError:",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
